+++ packages/console/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @hot-updater/console
 
+## 0.35.12
+
+### Patch Changes
+
+- 6e8b32e: Replace the semver dependency with verkit.
+- Updated dependencies [6e8b32e]
+  - @hot-updater/server@0.35.12
+  - @hot-updater/bsdiff@0.35.12
+
 ## 0.35.11
 
 ### Patch Changes